Move run_cli, fix optimizer of high res exp
