+++ bevdepth/exps/nuscenes/fusion/bev_depth_fusion_lss_r50_256x704_128x128_24e.py
@@ -4,7 +4,7 @@ import torch.nn.parallel
 import torch.utils.data
 import torch.utils.data.distributed
 
-from bevdepth.exps.nuscenes.base_cli import run_cli
+from bevdepth.exps.base_cli import run_cli
 from bevdepth.exps.nuscenes.base_exp import \
     BEVDepthLightningModel as BaseBEVDepthLightningModel
 from bevdepth.models.fusion_bev_depth import FusionBEVDepth
