+++ bevdepth/exps/nuscenes/fusion/bev_depth_fusion_lss_r50_256x704_128x128_24e_2key_trainval.py
@@ -1,7 +1,8 @@
 # Copyright (c) Megvii Inc. All rights reserved.
-from bevdepth.exps.nuscenes.base_cli import run_cli
-from bevdepth.exps.nuscenes.fusion.bev_depth_fusion_lss_r50_256x704_128x128_24e_2key import \
-    BEVDepthLightningModel as BaseBEVDepthLightningModel  # noqa
+from bevdepth.exps.base_cli import run_cli
+
+from .bev_depth_fusion_lss_r50_256x704_128x128_24e_2key import \
+    BEVDepthLightningModel as BaseBEVDepthLightningModel
 
 
 class BEVDepthLightningModel(BaseBEVDepthLightningModel):
