+++ bevdepth/exps/nuscenes/mv/bev_depth_lss_r50_256x704_128x128_24e_2key.py
@@ -22,7 +22,7 @@ bicycle 0.305   0.554   0.264   0.653   0.263   0.006
 traffic_cone    0.462   0.516   0.355   nan     nan     nan
 barrier 0.512   0.491   0.275   0.200   nan     nan
 """
-from bevdepth.exps.nuscenes.base_cli import run_cli
+from bevdepth.exps.base_cli import run_cli
 from bevdepth.exps.nuscenes.base_exp import \
     BEVDepthLightningModel as BaseBEVDepthLightningModel
 from bevdepth.models.base_bev_depth import BaseBEVDepth
